Completo; falta revisión

+++ webapp/delete.php
@@ -1,9 +1,16 @@
 <?php
+// Include config file
+require_once "config/configuracion.php";
+
+// Si no está logeado como empleado, lo llevamos a la página de login.
+if(!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true || !isset($_SESSION["num_identif"])){
+    header("location: login.php");
+    exit;
+}
+
 // Process delete operation after confirmation
 if(isset($_POST["id"]) && !empty($_POST["id"])){
-    // Include config file
-    require_once "config/configuracion.php";
-    
+
     // Prepare a delete statement
     $sql = "DELETE FROM vacuna WHERE id = ?";
